requirements.txt
# Core Dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interfaces
gradio>=4.0.0
streamlit>=1.28.0

# Machine Learning Utils
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter (Optional)
jupyter>=1.0.0
ipython>=8.0.0
